refactor(solver): move Atom property-access entry to a query_cache child module (size ratchet)
AgentName: StudioOpus
