+++ crates/tsz-solver/src/caches/query_cache.rs
@@ -601,37 +601,6 @@ impl<'a> QueryCache<'a> {
         self.property_cache.borrow_mut().insert(key, result);
     }
 
-    /// `Atom`-keyed property-access resolution shared by the `&str` and `Atom`
-    /// `QueryDatabase` entry points. The cache key is already `Atom`-based, so
-    /// callers holding an `Atom` skip the property-name re-hash entirely.
-    fn property_access_atom_with_options(
-        &self,
-        object_type: TypeId,
-        prop_atom: Atom,
-        no_unchecked_indexed_access: bool,
-    ) -> PropertyAccessResult {
-        // QueryCache doesn't have full TypeResolver capability, so use
-        // PropertyAccessEvaluator with the current QueryDatabase.
-        let exact_optional_property_types =
-            crate::caches::db::TypeCompilerOptions::exact_optional_property_types(self);
-        let key = (
-            object_type,
-            prop_atom,
-            no_unchecked_indexed_access,
-            exact_optional_property_types,
-        );
-        if let Some(result) = self.check_property_cache(key) {
-            return result;
-        }
-
-        let mut evaluator = crate::operations::property::PropertyAccessEvaluator::new(self);
-        evaluator.set_no_unchecked_indexed_access(no_unchecked_indexed_access);
-        evaluator.set_exact_optional_property_types(exact_optional_property_types);
-        let result = evaluator.resolve_property_access_atom(object_type, prop_atom);
-        self.insert_property_cache(key, result);
-        result
-    }
-
     fn check_element_access_cache(&self, key: ElementAccessTypeCacheKey) -> Option<TypeId> {
         self.element_access_cache.borrow().get(&key).copied()
     }
@@ -2042,3 +2011,7 @@ mod tests;
 // the 2000-line file-size cap; child modules retain private-field access.
 #[path = "query_cache_size.rs"]
 mod size;
+
+// `Atom`-keyed property access lives in a child module for the same reason.
+#[path = "query_cache_property.rs"]
+mod property;
